+++ api/chef_handler.go
@@ -6,39 +6,45 @@ import (
 	"math/rand"
 	"net/http"
 
+	db "github.com/aopontann/gin-sqlc/db/sqlc"
 	"github.com/aopontann/gin-sqlc/docs"
 	"github.com/gin-gonic/gin"
 	"github.com/go-playground/validator/v10"
 	"github.com/mattn/go-gimei"
 )
 
-func (s *Server) ListTrendRecipe(c *gin.Context) {
+type featuredChefResponse struct {
+	Data []db.FakeListFeaturedChefRow `json:"data"`
+}
+
+func (s *Server) ListFeaturedChef(c *gin.Context) {
 	const limit int32 = 10
-	list, err := s.q.FakeListTrendRecipe(context.Background(), limit)
+	var response featuredChefResponse
+	var err error
+	response.Data, err = s.q.FakeListFeaturedChef(context.Background(), limit)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
 
 	// ダミーデータ作成（本番では消す）
-	for i := 0; i < len(list); i++ {
-		list[i].Name = gimei.NewName().First.Katakana()
-		list[i].Introduction = gimei.NewAddress().String() + "。" + gimei.NewAddress().String() + "。" + gimei.NewAddress().String() + "。" + gimei.NewAddress().String() + "。" + gimei.NewAddress().String() + "。"
-		list[i].NumFav = rand.Int31n(1000)
-		list[i].Score = rand.Int31n(100)
+	for i := 0; i < len(response.Data); i++ {
+		response.Data[i].Name = gimei.NewName().String()
+		response.Data[i].NumFollower = rand.Int31n(1000)
+		response.Data[i].Score = rand.Int31n(100)
 	}
 
 	// レスポンス型バリデーション
 	validate := validator.New()
-	for i := 0; i < len(list); i++ {
-		// ListTrendRecipeRow型からJSONに変換
-		jsn, err := json.Marshal(list[i])
+	for i := 0; i < len(response.Data); i++ {
+		// ListFeaturedChefRow型からJSONに変換
+		jsn, err := json.Marshal(response.Data[i])
 		if err != nil {
 			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 			return
 		}
-		// JSONからdocs.TrendRecipe型に変換
-		var obj docs.TrendRecipe
+		// JSONからdocs.FeaturedChef型に変換
+		var obj docs.FeaturedChef
 		if err := json.Unmarshal(jsn, &obj); err != nil {
 			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 			return
@@ -51,5 +57,5 @@ func (s *Server) ListTrendRecipe(c *gin.Context) {
 		}
 	}
 
-	c.JSON(http.StatusOK, list)
+	c.JSON(http.StatusOK, response)
 }
